Issue 404: redirect working.
But need to figure out how to get the invitation success notification up for the user. Which either depends on work for #175, or showing a custom notification on the dashboard (which feels like really it should just be handled by #175) or to show the invitees on the dashboard to indicate success to the user.